fixed favorites icon and view
