+++ src/main/java/com/chowhound/chowhound/models/Truck.java
@@ -47,13 +47,18 @@ public class Truck {
     @Column(nullable = false, columnDefinition = "boolean default false")
     private boolean isVerified;
 
+    @Column
+    private String instagram;
+
+    @Column
+    private String facebook;
+
+    @Column
+    private String twitter;
+
     @OneToOne
     @JoinColumn(name = "user_id")
     private User user;
-
-//    SELECT *
-//    FROM  trucks AS t
-//    JOIN users AS u ON t.user_id = u.id;
 
     // relationship for images
     @OneToMany(mappedBy = "truck")
@@ -201,4 +206,15 @@ public class Truck {
         return reviews.stream().mapToInt(Review::getStars).average().orElse(0);
     }
 
+    public String getInstagram() { return instagram; }
+
+    public void setInstagram(String instagram) { this.instagram = instagram; }
+
+    public String getFacebook() { return facebook; }
+
+    public void setFacebook(String facebook) { this.facebook = facebook; }
+
+    public String getTwitter() { return twitter; }
+
+    public void setTwitter(String twitter) { this.twitter = twitter; }
 }
