+++ src/main/resources/templates/partials/partials.html
@@ -24,7 +24,7 @@
 <!--            <li class="nav-item"><a class="nav-link" href="/index">Find some chow</a></li>-->
             <li sec:authorize="isAuthenticated()" class="nav-item"><a class="nav-link" href="/trucks/create">Add a
                 Truck</a></li>
-            <li sec:authorize="isAuthenticated()" class="nav-item"><a class="nav-link" href="/trucks/my_favorites">My
+            <li sec:authorize="isAuthenticated()" class="nav-item"><a class="nav-link" href="/favorites">My
                 trucks</a></li>
             <!--            <li class="nav-item"><a class="nav-link" href="/welcome">Welcome</a></li>-->
         </ul>
