+++ src/main/resources/templates/trucks/create.html
@@ -53,6 +53,14 @@
             <input type="checkbox" id="owner" name="owner" value="true">
             <input id='ownerHidden' type='hidden' value='false' name='owner'>
 
+
+            <label for="instagram"></label>
+            <input type="text" id="instagram" name="instagram" th:field="*{instagram}">
+            <label for="facebook"></label>
+            <input type="text" id="facebook" name="facebook" th:field="*{facebook}">
+            <label for="twitter"></label>
+            <input type="text" id="twitter" name="twitter" th:field="*{twitter}">
+
             <br>
             <input value="Mark it dude!" type="submit"/>
